requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
data==0.4
decorator==5.1.1
frozenlist==1.4.0
funcsigs==1.0.2
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
Pillow==10.1.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.7
yarl==1.9.2